% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_launch.R
\name{launch_mapmixture}
\alias{launch_mapmixture}
\title{Launch Shiny App}
\usage{
launch_mapmixture(...)
}
\arguments{
\item{...}{additional arguments passed to \code{shiny::runApp}.}
}
\description{
Wrapper function used to start mapmixture shiny app.
}
\examples{
if (interactive()){
  launch_mapmixture(launch.browser = TRUE)
}
}
