% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_mapmixture.R
\name{mapmixture}
\alias{mapmixture}
\title{Map Pie Charts}
\usage{
mapmixture(
  admixture_df,
  coords_df,
  cluster_cols = NULL,
  cluster_names = NULL,
  boundary = NULL,
  crs = 4326,
  pie_size = 1,
  pie_border = 0.3,
  pie_opacity = 1,
  land_colour = "#d9d9d9",
  sea_colour = "#deebf7",
  expand = TRUE,
  arrow = TRUE,
  arrow_size = 1,
  arrow_position = "tl",
  scalebar = TRUE,
  scalebar_size = 1,
  scalebar_position = "tl",
  plot_title = "",
  plot_title_size = 15,
  axis_title_size = 12,
  axis_text_size = 10
)
}
\arguments{
\item{admixture_df}{a data.frame or tibble containing admixture data (see examples).}

\item{coords_df}{a data.frame or tibble containing coordinates data (see examples).}

\item{cluster_cols}{a character vector of colours the same length as the number of clusters.
If \code{NULL}, a blue-green palette is used.}

\item{cluster_names}{a character vector of names the same length as the number of clusters.
If \code{NULL}, the cluster column names are used.}

\item{boundary}{a named numeric vector defining the map bounding. e.g. \code{c(xmin=-15, xmax=15, ymin=30, ymax=50)}.
If \code{NULL}, a default bounding box is calculated.}

\item{crs}{coordinate reference system. Default is the WGS 84 - World Geodetic System 1984 (EPSG:\code{4326}).
See \code{?sf::st_crs} for details.}

\item{pie_size}{a numeric value of zero or greater.}

\item{pie_border}{a numeric value of zero or greater.}

\item{pie_opacity}{a numeric value of zero to one.}

\item{land_colour}{a string defining the colour of land.}

\item{sea_colour}{a string defining the colour of sea.}

\item{expand}{expand axes (\code{TRUE} or \code{FALSE}).}

\item{arrow}{show arrow (\code{TRUE} or \code{FALSE}).}

\item{arrow_size}{a numeric value of zero or greater.}

\item{arrow_position}{a string defining the position of the arrow (\code{"tl"}, \code{"tr"}, \code{"bl"}, \code{"br"}).}

\item{scalebar}{show scalebar (\code{TRUE} or \code{FALSE}).}

\item{scalebar_size}{a numeric value of zero or greater.}

\item{scalebar_position}{a string defining the position of the scalebar (\code{"tl"}, \code{"tr"}, \code{"bl"}, \code{"br"}).}

\item{plot_title}{a string defining the main title of the plot.}

\item{plot_title_size}{a numeric value of zero or greater.}

\item{axis_title_size}{a numeric value of zero or greater.}

\item{axis_text_size}{a numeric value of zero or greater.}
}
\value{
A ggplot object.
}
\description{
Plot admixture proportions as pie charts on a projected map.
In data sets where there are multiple individuals per site,
the function will calculate the mean average admixture proportion for each site.
}
\examples{
# Admixture Format 1
admixture1 <- data.frame(
  Site = c("London", "London", "Paris", "Paris", "Berlin", "Berlin",
   "Rome", "Rome", "Madrid", "Madrid"),
  Ind = c("London1", "London2", "Paris1", "Paris2", "Berlin1", "Berlin2",
   "Rome1", "Rome2", "Madrid1", "Madrid2"),
  Cluster1 = c(1.0, 0.9, 0.5, 0.5, 0.1, 0.1, 0, 0, 0, 0),
  Cluster2 = c(0, 0.10, 0.50, 0.40, 0.50, 0.40, 0.01, 0.01, 0.70, 0.80),
  Cluster3 = c(0, 0, 0, 0.10, 0.40, 0.50, 0.99, 0.99, 0.30, 0.20)
)

# Admixture Format 2
admixture2 <- data.frame(
  Site = c("London", "Paris", "Berlin", "Rome", "Madrid"),
  Ind = c("London", "Paris", "Berlin", "Rome", "Madrid"),
  Cluster1 = c(0.95, 0.5, 0.1, 0, 0),
  Cluster2 = c(0.05, 0.45, 0.45, 0.01, 0.75),
  Cluster3 = c(0, 0.05, 0.45, 0.99, 0.25)
)

# Admixture Format 3
admixture3 <- data.frame(
  Site = c("London", "Paris", "Berlin", "Rome", "Madrid"),
  Ind = c("London", "Paris", "Berlin", "Rome", "Madrid"),
  Cluster1 = c(1, 1, 0, 0, 0),
  Cluster2 = c(0, 0, 1, 0, 1),
  Cluster3 = c(0, 0, 0, 1, 0)
)

# Coordinates Format
coordinates <- data.frame(
  Site = c("London", "Paris", "Berlin", "Rome", "Madrid"),
  Lat = c(51.51, 48.85, 52.52, 41.90, 40.42),
  Lon = c(-0.12, 2.35, 13.40, 12.49, -3.70)
)

# Plot using default parameters
mapmixture(admixture1, coordinates)

# Plot using the ETRS89-extended / LAEA Europe coordinate reference system
mapmixture(admixture1, coordinates, crs = 3035)

# Plot using custom parameters
mapmixture(admixture1, coordinates,
  cluster_cols = c("blue","purple","green"),
  cluster_names = c("Group 1","Group 2", "Group 3"),
  crs = "+proj=merc +a=6378137 +b=6378137 +lat_ts=0 +lon_0=0 +x_0=0 +y_0=0 +units=m",
  boundary = c(xmin=-10, xmax=17, ymin=37, ymax=55),
  pie_size = 1.5,
  pie_border = 0.3,
  pie_opacity = 1,
  land_colour = "#d9d9d9",
  sea_colour = "#deebf7",
  expand = TRUE,
  arrow = TRUE,
  arrow_size = 1,
  arrow_position = "tl",
  scalebar = TRUE,
  scalebar_size = 1,
  scalebar_position = "tl",
  plot_title = "Mapmixture Figure",
  plot_title_size = 15,
  axis_title_size = 12,
  axis_text_size = 10
)
}
