% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_process_input_data.R
\name{read_input_data}
\alias{read_input_data}
\title{Read Input Data}
\usage{
read_input_data(file)
}
\arguments{
\item{file}{path to input file.}
}
\value{
A data.frame object.
}
\description{
Internal function used to read in data from admixture or coordinates input file.
}
\examples{
file_path <- system.file("extdata", "admixture1.csv", package = "mapmixture")
read_input_data(file_path)
}
\keyword{internal}
