% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_data_transformation.R
\name{transform_bbox}
\alias{transform_bbox}
\title{Transform Bounding Box Coordinates}
\usage{
transform_bbox(bbox, CRS)
}
\arguments{
\item{bbox}{named numeric vector of length four, with \code{xmin}, \code{xmax}, \code{ymin} and \code{ymax} values.}

\item{CRS}{numeric value representing an ESPG.}
}
\value{
A named numeric vector.
}
\description{
Internal function used to transform latitude and longitude coordinates
to a target coordinate reference system.
}
\examples{
bbox <- c(xmin = -3.70, xmax = 13.40, ymin = 40.42, ymax = 52.52)
transform_bbox(bbox, 3035)
}
\keyword{internal}
