% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{launch_mapmixture}
\alias{launch_mapmixture}
\title{Run Mapmixture Shiny App}
\usage{
launch_mapmixture(...)
}
\arguments{
\item{...}{additional arguments to pass to \code{shiny::runApp} call.}
}
\description{
Wrapper function used to start mapmixture shiny app.
}
