% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_launch.R
\name{launch_mapmixture}
\alias{launch_mapmixture}
\title{Launch Shiny App}
\usage{
launch_mapmixture(...)
}
\arguments{
\item{...}{additional arguments passed to \code{shiny::runApp()}.}
}
\value{
No return value.
}
\description{
Wrapper function used to start \code{mapmixture} interactive app.

App tested with the following package versions:

shiny v1.8.0,
shinyFeedback v0.4.0,
shinyjs v2.1.0,
shinyWidgets 0.8.4,
bslib 0.7.0,
colourpicker 1.3.0,
htmltools v0.5.8.1,
waiter 0.2.5.
}
\examples{
if (interactive()){
  launch_mapmixture(launch.browser = TRUE)
}
}
