% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_bar_chart.R
\name{structure_plot}
\alias{structure_plot}
\title{STRUCTURE Bar Chart}
\usage{
structure_plot(
  df,
  type = "structure",
  colours = NULL,
  site_dividers = TRUE,
  labels = "site",
  flip_axis = FALSE,
  facet_col = NULL,
  facet_row = NULL
)
}
\arguments{
\item{df}{a data frame (see examples).}
}
\value{
A ggplot object.
}
\description{
Internal function used to create a traditional STRUCTURE plot or
a facet bar chart based on admixture proportions.
}
\examples{
df <- data.frame(
  site = c("Berlin", "Berlin", "London", "London", "Madrid", "Madrid",
          "Paris", "Paris", "Rome", "Rome"),
  ind = c("Berlin1", "Berlin2", "London1", "London2", "Madrid1", "Madrid2",
          "Paris1", "Paris2", "Rome1", "Rome2"),
  cluster1 = c(0.1, 0.1, 1.0, 0.9, 0.0, 0.0, 0.5, 0.5, 0.0, 0.0),
  cluster2 = c(0.50, 0.40, 0.00, 0.10, 0.70, 0.80, 0.50, 0.40, 0.01, 0.01),
  cluster3 = c(0.40, 0.50, 0.00, 0.00, 0.30, 0.20, 0.00, 0.10, 0.99, 0.99)
)

df <- data.frame(
  site = c(rep("Site1", 30), rep("Site2", 30),
           rep("Site3", 30), rep("Site4", 30)),
  ind = c(paste0("site1_", 1:30), paste0("site2_", 1:30),
          paste0("site3_", 1:30), paste0("site4_", 1:30)),
  cluster1 = c(rep(0.5, 30), rep(0.25, 30), rep(0.05, 30), rep(0.05, 30)),
  cluster2 = c(rep(0.25, 30), rep(0.50, 30), rep(0.25, 30), rep(0.05, 30)),
  cluster3 = c(rep(0.20, 30), rep(0.20, 30), rep(0.20, 30), rep(0.50, 30)),
  cluster4 = c(rep(0.05, 30), rep(0.05, 30), rep(0.50, 30), rep(0.40, 30))
)

structure_plot(df, type = "structure")
}
